# frontend/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "PolicyPilot API",
    "description" : "Policies, rule set versions, decisions and their traces. The model proposes and explains, the rules engine decides, a person approves every policy change.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "This API"
  } ],
  "paths" : {
...
    "/api/v1/rulesets/{id}/versions/{no}/retrieval" : {
      "post" : {
        "tags" : [ "retrieval-controller" ],
        "summary" : "Retrieve the chunks of a version for a question, or the fixed not-covered sentence",
        "operationId" : "retrieve",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "no",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RetrievalRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "The fused chunks and their citations, or the not-covered sentence",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetrievalResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "The question is empty, too long or has a control character",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404" : {
            "description" : "No such rule set version in this sandbox",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409" : {
            "description" : "The version is not published, or its embedding is not READY",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503" : {
            "description" : "The embedding provider is unavailable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rulesets/{id}/versions/{no}" : {
      "get" : {
        "tags" : [ "ruleset-controller" ],
        "summary" : "A rule set version with its rules, findings and status",
        "operationId" : "version",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "no",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The version with its rules, findings and status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such rule set version in this sandbox",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AcknowledgementResponse" : {
        "type" : "object",
        "properties" : {
          "resolution" : {
            "type" : "string",
            "enum" : [ "rule_added", "flag_added", "interpretation" ]
          },
          "note" : {
            "type" : "string"
          },
          "at" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "at" ]
      },
      "FindingResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "severity" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "ruleIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "fieldNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "code", "fieldNames", "message", "path", "ruleIds", "severity" ]
      },
      "ReviewFindingResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "ambiguity", "conflict", "unsupported", "gap", "duplicate", "injection" ]
          },
          "severity" : {
            "type" : "string",
            "enum" : [ "error", "warning" ]
          },
          "ruleIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "paragraphIndexes" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "message" : {
            "type" : "string"
          },
          "suggestion" : {
            "type" : "string"
          },
          "confidence" : {
            "type" : "number",
            "format" : "double"
          },
          "blocking" : {
            "type" : "boolean"
          },
          "acknowledgement" : {
            "$ref" : "#/components/schemas/AcknowledgementResponse"
          }
        },
        "required" : [ "blocking", "confidence", "id", "kind", "message", "paragraphIndexes", "ruleIds", "severity", "suggestion" ]
      },
      "ReviewResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "DONE", "FAILED", "STALE" ]
          },
          "promptVersion" : {
            "type" : "string"
          },
          "findings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewFindingResponse"
            }
          },
          "coverage" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        },
        "required" : [ "coverage", "findings", "promptVersion", "status" ]
      },
      "VersionResponse" : {
        "type" : "object",
        "properties" : {
          "rulesetId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "domain" : {
            "type" : "string"
          },
          "protected" : {
            "type" : "boolean"
          },
          "forkedFromId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "versionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "versionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string"
          },
          "policyVersionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "parentVersionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "publishedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "publishedBy" : {
            "type" : "string"
          },
          "ruleSet" : {
            "description" : "The whole DSL document (Document 3)"
          },
          "findings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FindingResponse"
            }
          },
          "review" : {
            "$ref" : "#/components/schemas/ReviewResponse"
          }
        },
        "required" : [ "domain", "findings", "name", "policyVersionId", "protected", "ruleSet", "rulesetId", "status", "versionId", "versionNo" ]
      },
      "ErrorDetail" : {
        "type" : "object",
        "properties" : {
          "path" : {
            "type" : "string"
          },
          "problem" : {
            "type" : "string"
          }
        },
        "required" : [ "path", "problem" ]
      },
      "ErrorEnvelope" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "details" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDetail"
            }
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "details", "message", "traceId" ]
      },
      "RetrievalRequest" : {
        "type" : "object",
        "properties" : {
          "question" : {
            "type" : "string"
          }
        }
      },
      "Chunk" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          },
          "score" : {
            "type" : "number",
            "format" : "double"
          },
          "cosine" : {
            "type" : "number",
            "format" : "double"
          },
          "vectorRank" : {
            "type" : "integer",
            "format" : "int32"
          },
          "lexicalRank" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Cite" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "paragraph" : {
            "type" : "integer",
            "format" : "int32"
          },
          "ruleId" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          }
        }
      },
      "RetrievalResponse" : {
        "type" : "object",
        "properties" : {
          "covered" : {
            "type" : "boolean"
          },
          "bestCosine" : {
            "type" : "number",
            "format" : "double"
          },
          "chunks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Chunk"
            }
          },
          "citations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Cite"
            }
          },
          "notCovered" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}